delete vesktop
kinda dumb that i have to go to github and manually remove files that i deleted on my local machine before i can commit but whatever.
vesktop doesnt work properly if this file is symlinked so im just not gonna add it to my dotfiles repo. it only takes me like a minute to reenable the plugins anyways.